Move it into a separate module.
